Added penTasks

+++ src/Consts.java
@@ -14,8 +14,8 @@ public class Consts {
   public final static int DEFAULT_SERVER_PORT = 20002;
   public final static int DEFAULT_WIDTH = 10;
 
-  public final static int DEFAULT_PAINTMODE = MODE_NORMAL;
+  public final static byte DEFAULT_PAINTMODE = MODE_NORMAL;
 
-  public final static int DEFAULT_FILLMODE = NOFILL;
+  public final static byte DEFAULT_FILLMODE = NOFILL;
   public final static Color DEFAULT_COLOR = Colors.BLACK;
 }
